Move user-agent header provider to its own package
Also makes the provider generic to any header
